Converted PKSanc to new API structure
