+++ app/Enum/PKSanc/ContributionStrings.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Enum\PKSanc;
+
+class ContributionStrings
+{
+    public const ROMHACK_ADDED = 'The romhack was succesfully added.';
+}
