+++ app/Http/Controllers/Modules/PKSanc/PKSancDepositController.php
@@ -37,50 +37,6 @@ class PKSancDepositController extends Controller
     }
 
     /**
-     * Attempt to stage a deposit
-     * @return View|RedirectResponse Returns a View of the page or a redirection response
-     */
-    public function stageDepositAttempt(Request $request): View|RedirectResponse
-    {
-        $user = Auth::user();
-        $validator = Validator::make($request->all(), [
-            'name' => 'required|string|min:4|max:255',
-            'csv' => 'required|mimes:csv,txt|max:480',
-            'game' => 'required|exists:pksanc__game,game'
-        ]);
-
-        if ($validator->fails()) {
-            return back()
-                ->withInput($request->all())
-                ->withErrors($validator);
-        }
-
-        $game = Game::where('game', $request->game)->first();
-        $date = Carbon::now()->unix();
-        $filename = sprintf('%s (%s).csv', $game->name, $date);
-
-        $path = sprintf(
-            StoragePaths::csv,
-            $user->uuid,
-            $request->name,
-        );
-
-        Storage::putFileAs($path, $request->file('csv'), $filename);
-
-        $csv = new ImportCsv();
-        $csv->csv = $filename;
-        $csv->game = $game->game;
-        $csv->name = $request->name;
-        $csv->uploader_uuid = $user->uuid;
-        $csv->version = 1;
-        $csv->save();
-
-        $this->depositService->stageImport($csv);
-
-        return redirect(route('pksanc.deposit.stage.show', $csv->uuid));
-    }
-
-    /**
      * Show a staging attempt
      * @return View Returns a View of the page
      */
