+++ resources/ts/components/datalists/DataSelect.ts
@@ -2,18 +2,17 @@ import {DataSelect as LaravelDataselect} from "@lati111/laravel_datatables/src/T
 import {getData as fetchGet, postData as fetchPost} from "../../ajax";
 
 /** @inheritDoc */
-
 export class DataSelect extends LaravelDataselect {
     /** @inheritDoc */
     async fetchData(url: string): Promise<any> {
         const response = await fetchGet(url);
-        return response;
+        return response.data;
     }
 
     /** @inheritDoc */
     async postData(url: string, parameters: FormData): Promise<any> {
         const response = await fetchPost(url, parameters);
-        //response?.announce();
-        return response;
+        response?.announce();
+        return response?.data;
     }
 }
