+++ resources/ts/modules/pksanc/deposit.ts
@@ -43,8 +43,11 @@ async function saveRomhack() {
     formdata.append('name', nameInput.value);
     formdata.append('original_game', originalGameSelector.value);
 
-    await postData('/pksanc/romhacks/add', formdata);
-    closeModal(romhackModalId);
+    const response = await postData('/pksanc/romhacks/add', formdata);
+    response?.announce();
+    if (response?.ok) {
+        closeModal(romhackModalId);
+    }
 }
 
 (<any>window).init = init;
