+++ routes/web.php
@@ -76,16 +76,6 @@ Route::prefix('/pksanc')
         });
     });
 
-    // api calls
-    Route::post('/romhacks/add', [PKSancContributionController::class, 'addRomhack'])
-        ->name('pksanc.games.romhacks.add');
-
-    Route::post('/pokedex/mark', [PKSancPokdexController::class, 'setPokedexMarking'])
-        ->name('pksanc.pokedex.mark');
-
-    Route::post('/pokedex/unmark', [PKSancPokdexController::class, 'setPokedexMarking'])
-        ->name('pksanc.pokedex.unmark');
-
     // data providers
     Route::prefix('/data/overview')->group(function() {
         Route::get('/', [PKSancOverviewCardList::class, 'data'])
